backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.2.0
pandas==2.3.3
psycopg[binary]==3.1.19
